* Update vc8 project file * Add LUA support to vc8 project file * Change LUA_NUMBER in the source & remove   LUA_NUMBER=float preprocessor directive.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1921 37977431-3df6-0310-b722-df95706aa16b

+++ lua/lua/src/lib/streflop_gateway.cpp
@@ -10,7 +10,7 @@ See lmathlib.c for details.
 
 // stolen from lua.h
 #ifndef LUA_NUMBER
-typedef double lua_Number;
+typedef float lua_Number;
 #else
 typedef LUA_NUMBER lua_Number;
 #endif
